+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/CreateBitsnBobs.java
@@ -7,7 +7,12 @@ import com.kipti.bnb.content.kinetics.cogwheel_chain.types.BnbCogwheelChainTypes
 import com.kipti.bnb.network.BnbPackets;
 import com.kipti.bnb.registry.azimuth.BnbBehaviourApplicators;
 import com.kipti.bnb.registry.compat.BnbCreateStresses;
-import com.kipti.bnb.registry.content.*;
+import com.kipti.bnb.registry.content.BnbAdvancements;
+import com.kipti.bnb.registry.content.BnbBlockEntities;
+import com.kipti.bnb.registry.content.BnbBlocksBootstrap;
+import com.kipti.bnb.registry.content.BnbContraptionTypes;
+import com.kipti.bnb.registry.content.BnbEntityTypes;
+import com.kipti.bnb.registry.content.BnbItems;
 import com.kipti.bnb.registry.core.BnbConfigs;
 import com.kipti.bnb.registry.core.BnbDataComponents;
 import com.kipti.bnb.registry.core.BnbTags;