+++ mods/Create-Bits-n-Bobs/src/main/java/com/kipti/bnb/content/kinetics/cogwheel_chain/graph/CogwheelChainCandidate.java
@@ -59,7 +59,7 @@ public record CogwheelChainCandidate(Direction.Axis axis, boolean isLarge, boole
     }
 
     public static @Nullable CogwheelChainCandidate getForBlock(final BlockState state) {
-        if (!isValidCandidate(state))
+        if (state == null || !isValidCandidate(state))
             return null;
         return new CogwheelChainCandidate(getAxis(state), isLargeCogwheel(state), hasSmallCogwheelOffset(state));
     }